Remove SFML RenderWindow stuff so we always use wxGLCanvas (and FTGL)
This simplifies things a fair bit from a code perspective, at the cost of 1 extra dependency in Windows (FTGL), but that is very easy to setup now via vcpkg.

The only real advantage to using the SFML RenderWindow was using SFML's text rendering (which allows text outline among other things), however I plan to move over to 'modern' opengl in the near future, which would mean we couldn't use the SFML rendering features anyway.
