+++ msvc/SLADE.vcxproj.filters
@@ -777,12 +777,6 @@
     <ClCompile Include="..\src\UI\Controls\STabCtrl.cpp">
       <Filter>UI\Controls</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\OpenGL\DrawingSFML.cpp">
-      <Filter>OpenGL</Filter>
-    </ClCompile>
-    <ClCompile Include="..\src\OpenGL\DrawingFTGL.cpp">
-      <Filter>OpenGL</Filter>
-    </ClCompile>
     <ClCompile Include="..\src\SLADEMap\MapObjectCollection.cpp">
       <Filter>SLADEMap</Filter>
     </ClCompile>
