+++ src/Application/App.h
@@ -64,7 +64,6 @@ namespace app
 	};
 	Platform      platform();
 	bool          useWebView();
-	bool          useSFMLRenderWindow();
 	const string& iconFile();
 	bool          isWin64Build();
