+++ src/OpenGL/Drawing.cpp
@@ -31,13 +31,14 @@
 // -----------------------------------------------------------------------------
 #include "Main.h"
 #include "Drawing.h"
+#include "App.h"
 #include "Archive/ArchiveManager.h"
 #include "GLTexture.h"
-#include "General/Misc.h"
 #include "General/UI.h"
 #include "OpenGL.h"
 #include "Utility/MathStuff.h"
 #include "Utility/StringUtils.h"
+#include <FTGL/ftgl.h>
 
 #ifdef __WXGTK3__
 #include <gtk-3.0/gtk/gtk.h>
@@ -64,7 +65,13 @@ CVAR(Int, gl_font_size, 12, CVar::Flag::Save)
 namespace slade::drawing
 {
 double  text_outline_width = 0;
-ColRGBA outline_colour     = ColRGBA::BLACK;
+ColRGBA outline_colour     = ColRGBA{ 0, 0, 0, 255 };
+FTFont* font_normal        = nullptr;
+FTFont* font_condensed     = nullptr;
+FTFont* font_bold          = nullptr;
+FTFont* font_boldcondensed = nullptr;
+FTFont* font_mono          = nullptr;
+FTFont* font_small         = nullptr;
 }; // namespace slade::drawing
 
 
@@ -73,6 +80,182 @@ ColRGBA outline_colour     = ColRGBA::BLACK;
 // Drawing Namespace Functions
 //
 // -----------------------------------------------------------------------------
+namespace slade::drawing
+{
+// -----------------------------------------------------------------------------
+// Loads all needed fonts for rendering
+// -----------------------------------------------------------------------------
+int initFonts()
+{
+	// --- Load general fonts ---
+	int ret = 0;
+
+	if (font_normal)
+	{
+		delete font_normal;
+		font_normal = nullptr;
+	}
+	if (font_condensed)
+	{
+		delete font_condensed;
+		font_condensed = nullptr;
+	}
+	if (font_bold)
+	{
+		delete font_bold;
+		font_bold = nullptr;
+	}
+	if (font_boldcondensed)
+	{
+		delete font_boldcondensed;
+		font_boldcondensed = nullptr;
+	}
+	if (font_mono)
+	{
+		delete font_mono;
+		font_mono = nullptr;
+	}
+	if (font_small)
+	{
+		delete font_small;
+		font_small = nullptr;
+	}
+
+	// Normal
+	auto entry = app::archiveManager().programResourceArchive()->entryAtPath("fonts/dejavu_sans.ttf");
+	if (entry)
+	{
+		font_normal = new FTTextureFont(entry->rawData(), entry->size());
+		font_normal->FaceSize(ui::scalePx(gl_font_size));
+
+		// Check it loaded ok
+		if (font_normal->Error())
+		{
+			delete font_normal;
+			font_normal = nullptr;
+		}
+		else
+			++ret;
+	}
+
+	// Condensed
+	entry = app::archiveManager().programResourceArchive()->entryAtPath("fonts/dejavu_sans_c.ttf");
+	if (entry)
+	{
+		font_condensed = new FTTextureFont(entry->rawData(), entry->size());
+		font_condensed->FaceSize(ui::scalePx(gl_font_size));
+
+		// Check it loaded ok
+		if (font_condensed->Error())
+		{
+			delete font_condensed;
+			font_condensed = nullptr;
+		}
+		else
+			++ret;
+	}
+
+	// Bold
+	entry = app::archiveManager().programResourceArchive()->entryAtPath("fonts/dejavu_sans_b.ttf");
+	if (entry)
+	{
+		font_bold = new FTTextureFont(entry->rawData(), entry->size());
+		font_bold->FaceSize(ui::scalePx(gl_font_size));
+
+		// Check it loaded ok
+		if (font_bold->Error())
+		{
+			delete font_bold;
+			font_bold = nullptr;
+		}
+		else
+			++ret;
+	}
+
+	// Condensed bold
+	entry = app::archiveManager().programResourceArchive()->entryAtPath("fonts/dejavu_sans_cb.ttf");
+	if (entry)
+	{
+		font_boldcondensed = new FTTextureFont(entry->rawData(), entry->size());
+		font_boldcondensed->FaceSize(ui::scalePx(gl_font_size));
+
+		// Check it loaded ok
+		if (font_boldcondensed->Error())
+		{
+			delete font_boldcondensed;
+			font_boldcondensed = nullptr;
+		}
+		else
+			++ret;
+	}
+
+	// Monospace
+	entry = app::archiveManager().programResourceArchive()->entryAtPath("fonts/dejavu_mono.ttf");
+	if (entry)
+	{
+		font_mono = new FTTextureFont(entry->rawData(), entry->size());
+		font_mono->FaceSize(ui::scalePx(gl_font_size));
+
+		// Check it loaded ok
+		if (font_mono->Error())
+		{
+			delete font_mono;
+			font_mono = nullptr;
+		}
+		else
+			++ret;
+	}
+
+	// Small
+	entry = app::archiveManager().programResourceArchive()->entryAtPath("fonts/dejavu_sans.ttf");
+	if (entry)
+	{
+		font_small = new FTTextureFont(entry->rawData(), entry->size());
+		font_small->FaceSize((ui::scalePx(gl_font_size) * 0.6) + 1);
+
+		// Check it loaded ok
+		if (font_small->Error())
+		{
+			delete font_small;
+			font_small = nullptr;
+		}
+		else
+			++ret;
+	}
+
+	return ret;
+}
+
+// -----------------------------------------------------------------------------
+// Cleans up all created fonts
+// -----------------------------------------------------------------------------
+void cleanupFonts()
+{
+	delete font_normal;
+	delete font_condensed;
+	delete font_bold;
+	delete font_boldcondensed;
+	delete font_mono;
+	delete font_small;
+}
+
+// -----------------------------------------------------------------------------
+// Returns the requested [font]
+// -----------------------------------------------------------------------------
+FTFont* getFont(Font font)
+{
+	switch (font)
+	{
+	case Font::Normal: return font_normal;
+	case Font::Condensed: return font_condensed;
+	case Font::Bold: return font_bold;
+	case Font::BoldCondensed: return font_boldcondensed;
+	case Font::Monospace: return font_mono;
+	case Font::Small: return font_small;
+	default: return font_normal;
+	}
+}
+} // namespace slade::drawing
 
 
 // -----------------------------------------------------------------------------
@@ -143,7 +326,13 @@ void drawing::drawLineTabbed(Vec2d start, Vec2d end, double tab, double tab_max)
 // Draws a line from [p1] to [p2] with an arrowhead at the [p1] end.
 // If [twoway] is true, an arrowhead is also drawn at the [p2] end
 // -----------------------------------------------------------------------------
-void drawing::drawArrow(Vec2d p1, Vec2d p2, const ColRGBA& color, bool twoway, double arrowhead_angle, double arrowhead_length)
+void drawing::drawArrow(
+	Vec2d          p1,
+	Vec2d          p2,
+	const ColRGBA& color,
+	bool           twoway,
+	double         arrowhead_angle,
+	double         arrowhead_length)
 {
 	Vec2d  a1l, a1r, a2l, a2r;
 	Vec2d  vector = p1 - p2;
@@ -492,6 +681,81 @@ void drawing::drawTextureWithin(
 }
 
 // -----------------------------------------------------------------------------
+// Draws [text] at [x,y]. If [bounds] is not null, the bounding coordinates of
+// the rendered text string are written to it.
+// -----------------------------------------------------------------------------
+void drawing::drawText(const string& text, int x, int y, ColRGBA colour, Font font, Align alignment, Rectd* bounds)
+{
+	// Get desired font
+	auto ftgl_font = getFont(font);
+
+	// If FTGL font is invalid, do nothing
+	if (!ftgl_font)
+		return;
+
+	// Setup alignment
+	auto  bbox   = ftgl_font->BBox(text.c_str(), -1);
+	int   xpos   = x;
+	int   ypos   = y;
+	float width  = bbox.Upper().X() - bbox.Lower().X();
+	float height = ftgl_font->LineHeight();
+	if (alignment != Align::Left)
+	{
+		if (alignment == Align::Center)
+			xpos -= math::round(width * 0.5);
+		else
+			xpos -= width;
+	}
+
+	// Set bounds rect
+	if (bounds)
+	{
+		bbox = ftgl_font->BBox(text.c_str(), -1, FTPoint(xpos, ypos));
+		bounds->set(bbox.Lower().X(), bbox.Lower().Y(), bbox.Upper().X(), bbox.Lower().Y() + height);
+	}
+
+	// Draw the string
+	glPushMatrix();
+	glTranslatef(xpos, ypos + ftgl_font->FaceSize(), 0.0f);
+	glTranslatef(-0.375f, -0.375f, 0);
+	glScalef(1.0f, -1.0f, 1.0f);
+	if (text_outline_width > 0)
+	{
+		// Draw outline if set
+		gl::setColour(outline_colour);
+		glTranslatef(-2.0f, -1.0f, 0.0f);
+		ftgl_font->Render(text.c_str(), -1);
+		glTranslatef(0.0f, 2.0f, 0.0f);
+		ftgl_font->Render(text.c_str(), -1);
+		glTranslatef(4.0f, 0.0f, 0.0f);
+		ftgl_font->Render(text.c_str(), -1);
+		glTranslatef(0.0f, -2.0f, 0.0f);
+		ftgl_font->Render(text.c_str(), -1);
+		glTranslatef(-2.0f, 1.0f, 0.0f);
+	}
+	gl::setColour(colour);
+	ftgl_font->Render(text.c_str(), -1);
+	glPopMatrix();
+}
+
+// -----------------------------------------------------------------------------
+// Returns the width and height of [text] when drawn with [font]
+// -----------------------------------------------------------------------------
+Vec2d drawing::textExtents(const string& text, Font font)
+{
+	// Get desired font
+	auto ftgl_font = getFont(font);
+
+	// If FTGL font is invalid, return empty
+	if (!ftgl_font)
+		return { 0, 0 };
+
+	// Return width and height of text
+	auto bbox = ftgl_font->BBox(text.c_str(), -1);
+	return { bbox.Upper().X() - bbox.Lower().X(), ftgl_font->LineHeight() };
+}
+
+// -----------------------------------------------------------------------------
 // Sets the [thickness] and [colour] of the outline to use when drawing text
 // -----------------------------------------------------------------------------
 void drawing::setTextOutline(double thickness, const ColRGBA& colour)
@@ -639,9 +903,7 @@ wxColour drawing::darkColour(const wxColour& colour, float percent)
 // TextBox class constructor
 // -----------------------------------------------------------------------------
 TextBox::TextBox(string_view text, drawing::Font font, int width, int line_height) :
-	font_{ font },
-	width_{ width },
-	line_height_{ line_height }
+	font_{ font }, width_{ width }, line_height_{ line_height }
 {
 	setText(text);
 }
@@ -756,11 +1018,9 @@ void TextBox::setSize(int width)
 // -----------------------------------------------------------------------------
 // Draws the text box
 // -----------------------------------------------------------------------------
-void TextBox::draw(int x, int y, const ColRGBA& colour, drawing::Align alignment)
+void TextBox::draw(int x, int y, const ColRGBA& colour, drawing::Align alignment) const
 {
 	Rectd b;
-	drawing::enableTextStateReset(false);
-	drawing::setTextState(true);
 	for (const auto& line : lines_)
 	{
 		drawText(line, x, y, colour, font_, alignment, &b);
@@ -770,6 +1030,4 @@ void TextBox::draw(int x, int y, const ColRGBA& colour, drawing::Align alignment
 		else
 			y += line_height_;
 	}
-	drawing::enableTextStateReset(true);
-	drawing::setTextState(false);
 }
